requirements.txt
Django==5.2.1
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.31.0 